Generate stars globally on server side
This started as an attempt to make the menu background consistent with
the game (which has succeeded), but has had the knock on effect of
making all players see the exact same background.

The background is regenerated once an hour on the server and gets
injected into the menu and game HTML files.

This can make testing HTML changes a bit more cumbersome, but they can
still be accessed statically by editing the URL.