Lay out some server-side game logic refactor
Just the basic building blocks we'll need. The server will have to
initalise and track all objects in the game and then the clients will
have to render them based on that information is distributed from the
server.

Clients will have to run their own simulation too which defers to the
values from the server in cases where they differ.